feat: move BaseCallFilter to logion_config.
logion-network/logion-internal#1283
